requirements.txt
dash>=2.14.0
plotly>=5.18.0
pandas>=2.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
